feat: board controller
