+++ src/main.js
@@ -1,97 +1,23 @@
 import BoardComponent from "./components/board.js";
+import BoardController from "./controllers/board.js";
 import FilterComponent from "./components/filter.js";
-import LoadMoreButtonComponent from "./components/load-more-button.js";
-import TaskEditComponent from "./components/task-edit.js";
-import TaskComponent from "./components/task.js";
-import TasksComponent from "./components/tasks.js";
-import NoTasksComponent from "./components/no-tasks.js";
 import SiteMenuComponent from "./components/menu.js";
-import SortComponent from "./components/sorting.js";
-import {generateTasks} from './mocks/task.js';
-import {generateFilters} from './mocks/filter.js';
-import {InsertionPosition, render, remove, replace} from "./utils/render.js";
+import {generateTasks} from "./mocks/task.js";
+import {generateFilters} from "./mocks/filter.js";
+import {render, InsertionPosition} from "./utils/render.js";
 
 const TASKS_COUNT = 22;
-const SHOWING_TASKS_COUNT_ON_START = 8;
-const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
 
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
 const randomTasks = generateTasks(TASKS_COUNT);
 const randomFilters = generateFilters();
 
-const renderTask = (taskListElement, task) => {
-  const replaceTaskToEdit = () => {
-    replace(taskEditComponent, taskComponent);
-  };
-
-  const replaceEditToTask = () => {
-    replace(taskComponent, taskEditComponent);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      replaceEditToTask();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const taskComponent = new TaskComponent(task);
-  taskComponent.setEditButtonClickHandler(() => {
-    replaceTaskToEdit();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  const taskEditComponent = new TaskEditComponent(task);
-  taskEditComponent.setSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceEditToTask();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-
-  render(taskListElement, taskComponent, InsertionPosition.BEFOREEND);
-};
-
-const renderBoard = (boardComponent, tasks) => {
-  const isAllTasksArchived = tasks.every((task) => task.isArchive);
-
-  if (isAllTasksArchived) {
-    render(boardComponent.getElement(), new NoTasksComponent(), InsertionPosition.BEFOREEND);
-    return;
-  }
-
-  render(boardComponent.getElement(), new SortComponent(), InsertionPosition.BEFOREEND);
-  render(boardComponent.getElement(), new TasksComponent(), InsertionPosition.BEFOREEND);
-
-  const taskListElement = boardComponent.getElement().querySelector(`.board__tasks`);
-
-  let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
-  tasks.slice(0, showingTasksCount)
-    .forEach((task) => {
-      renderTask(taskListElement, task);
-    });
-
-  const loadMoreButtonComponent = new LoadMoreButtonComponent();
-  render(boardComponent.getElement(), loadMoreButtonComponent, InsertionPosition.BEFOREEND);
-
-  loadMoreButtonComponent.setClickHandler(() => {
-    const prevTasksCount = showingTasksCount;
-    showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
-
-    tasks.slice(prevTasksCount, showingTasksCount)
-      .forEach((task) => renderTask(taskListElement, task));
-
-    if (showingTasksCount >= tasks.length) {
-      remove(loadMoreButtonComponent);
-    }
-  });
-};
-
 render(siteHeaderElement, new SiteMenuComponent(), InsertionPosition.BEFOREEND);
 render(siteMainElement, new FilterComponent(randomFilters), InsertionPosition.BEFOREEND);
 
 const boardComponent = new BoardComponent();
+const boardController = new BoardController(boardComponent);
+
 render(siteMainElement, boardComponent, InsertionPosition.BEFOREEND);
-renderBoard(boardComponent, randomTasks);
+boardController.render(randomTasks);
